requirements.txt
gradio
gpt4all
nomic
requests
python-dotenv